refactor: change project structure, separating in responsabilities
