+++ frontend/app/tsconfig.app.json
@@ -23,5 +23,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src"]
+  "include": ["src", "src/features/products/components/ProductModal.tsx"]
 }
